pyproject.toml
[project]
name = "backend-aguas-cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio-pool>=0.6.0",
    "azure-functions>=1.23.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "colorama>=0.4.6",
    "pandas>=3.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "ruff>=0.12.12",
    "tabulate>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
